+++ packages/convert/app/components/FileAvailable.tsx
@@ -6,12 +6,12 @@ import {defaultRotateOrMirorState, RotateOrMirrorState} from '~/lib/default-ui';
 import {formatBytes} from '~/lib/format-bytes';
 import {useThumbnail} from '~/lib/use-thumbnail';
 import {RouteAction} from '~/seo';
+import {BackButton} from './BackButton';
 import ConvertUI from './ConvertUi';
 import {Footer} from './Footer';
 import {Probe} from './Probe';
 import {VideoThumbnailRef} from './VideoThumbnail';
 import {getBrightnessOfFrame} from './get-brightness-of-frame';
-import {Button} from './ui/button';
 import {useProbe} from './use-probe';
 
 const idealBrightness = 0.8;
@@ -24,10 +24,6 @@ export const FileAvailable: React.FC<{
 	const [probeDetails, setProbeDetails] = useState(
 		() => routeAction.type === 'generic-probe',
 	);
-
-	const clear = useCallback(() => {
-		setSrc(null);
-	}, [setSrc]);
 
 	const bestBrightness = useRef<number | null>(null);
 
@@ -99,22 +95,7 @@ export const FileAvailable: React.FC<{
 		<div>
 			<div className="overflow-y-auto w-full lg:flex lg:justify-center pt-6 pb-10 px-4 bg-slate-50 min-h-[100vh]">
 				<div>
-					<div className="block">
-						<Button variant="link" onClick={clear}>
-							<svg
-								xmlns="http://www.w3.org/2000/svg"
-								viewBox="0 0 448 512"
-								style={{height: 16}}
-							>
-								<path
-									fill="currentcolor"
-									d="M18.2 273l-17-17 17-17L171.8 85.4l17-17 33.9 33.9-17 17L93.1 232 424 232l24 0 0 48-24 0L93.1 280 205.8 392.6l17 17-33.9 33.9-17-17L18.2 273z"
-								/>
-							</svg>
-							<div className="w-2" />
-							Choose another file
-						</Button>
-					</div>
+					<BackButton setSrc={setSrc} />
 					<div className="lg:inline-flex lg:flex-row items-start">
 						<Probe
 							thumbnailError={err}
